requirements.txt
python-dotenv
certifi
databricks-sql-connector
azure-functions
azure-identity
azure-keyvault-secrets